refactor key_format from programming expression into a helper and use in ref guides
