+++ dashboard/app/models/programming_expression.rb
@@ -20,6 +20,7 @@
 #  programming_environment_key                                  (programming_environment_id,key) UNIQUE
 #
 class ProgrammingExpression < ApplicationRecord
+  include CurriculumHelper
   include SerializedProperties
 
   belongs_to :programming_environment
@@ -43,26 +44,6 @@ class ProgrammingExpression < ApplicationRecord
     video_key
     block_name
   )
-
-  def key_format
-    if key.blank?
-      errors.add(:base, 'Key must not be blank')
-      return false
-    end
-
-    if key[0] == '.' || key[-1] == '.'
-      errors.add(:base, 'Key cannot start or end with period')
-      return false
-    end
-
-    key_char_re = /[A-Za-z0-9\-\_\.]/
-    key_re = /\A#{key_char_re}+\Z/
-    unless key_re.match?(key)
-      errors.add(:base, "must only be letters, numbers, dashes, underscores, and periods. Got ${key}")
-      return false
-    end
-    return true
-  end
 
   def self.properties_from_file(path, content)
     expression_config = JSON.parse(content)
